QUIPU-458 fix after review
