+++ src/modules/new-liquidity/pages/list/new-liquidity-list-page.module.scss
@@ -5,7 +5,6 @@
 }
 
 .lastElement {
-  width: 100%;
   height: 1px;
   background-color: transparent;
 }
